E2E things working

+++ api/pkg/storage/disk.go
@@ -31,7 +31,9 @@ func (d *Disk) Upload(filePath, fileType, fileID string) (string, error) {
 	if err != nil {
 		return "", fmt.Errorf("error uploading file to disk as cannot copy contents: %w", err)
 	}
-	defer os.Remove(filePath)
+	if fileType != "previews" {
+		defer os.Remove(filePath)
+	}
 	return result, nil
 }
